Output a DOCTYPE to put the browser in the right mode
If krumo() is the first output it may be outside the DOCTYPE scope
and end up putting the browser in Quirks mode. This may affect later
rendering of the page and cause unforseen challenges.

This adds a DOCTYPE to the Krumo output and may result in multiple
DOCTYPEs in your HTML. This isn't 100% ideal/legit but it's better
than putting the browser in quirks mode